changes to enable DTO

+++ games-parent/external-api/src/main/java/br/edu/infnet/games/clients/GameDBFeignClient.java
@@ -14,6 +14,6 @@ public interface GameDBFeignClient {
     PlatformResponse findPlatformsByName(@RequestParam("name") String name);
 
     @GetMapping("/v1/Games/ByPlatformID?apikey=${api.theGamesDb.key}")
-    PlatformGamesResponse findGamesByPlatformID(@RequestParam("id") Long id);
+    PlatformGamesResponse findGamesByPlatformID(@RequestParam("id") Integer id);
 
 }
